Prioritize editorial human handoff with YT-first action candidates
