+++ docs/runbooks/README.md
@@ -1,0 +1,31 @@
+# Runbooks index
+
+Guía rápida para distinguir qué documento usar según objetivo.
+
+## Active operations
+
+- `pr5-minimal-autonomous-loop.md` — ruta operativa alive-first por lanes.
+- `pr5-minimal-pr-plan.md` — plan mínimo de ejecución PR5.
+
+## Consolidation and pruning
+
+- `pr5-pruning-diagnostic-and-phased-plan.md` — pruning guiado por runtime con guardrails.
+- `pr5-observability-indexes-prep.md` — preparación de índices de observabilidad compacta.
+- `pr6-wrapper-surface-inventory.md` — inventario de wrappers/helpers y shortlist canónica de comandos operativos.
+- `pr7-editorial-handoff-5day-sprint-plan.md` — plan de sprint de 5 días para hacer dominante el handoff editorial humano (YT-priority).
+
+## Ownership and migration records
+
+- `pr4a-news-acquire-ownership.md`
+- `pr4b-news-editorial-ownership.md`
+- `pr4c-news-acquire-code-migration.md`
+- `pr4d-news-editorial-migration.md`
+- `pr4e-news-enrich-migration.md`
+
+## Historical runtime evidence
+
+- `runtime-evidence-20260101T10.md`
+- `runtime-evidence-20260313T15.md`
+- `pr1b-runtime-validation-recovery-plan.md`
+
+> Criterio práctico: si necesitás correr el sistema hoy, empezá por `pr5-minimal-autonomous-loop.md`.
