+++ tests/test_build_editorial_access_indexes.py
@@ -47,6 +47,7 @@ def test_build_editorial_access_indexes_with_partial_data(tmp_path: Path):
                 "schema_name": "news_piece_brief.v1",
                 "digest_id_hour": digest,
                 "brief_id": "npb_1",
+                "format_candidates": ["yt_script", "article"],
             }
         ],
     )
@@ -88,6 +89,13 @@ def test_build_editorial_access_indexes_with_partial_data(tmp_path: Path):
     assert payload["metrics"]["fallback_legacy_count"] == 1
     assert payload["metrics"]["schema_failures"] == 1
     assert payload["status"] == "degraded"
+    assert payload["human_handoff"]["status"] == "needs-attention"
+    assert len(payload["human_handoff"]["latest_briefs"]) == 1
+    assert len(payload["human_handoff"]["latest_article_drafts"]) == 1
+    assert payload["human_handoff"]["latest_yt_script_drafts"] == []
+    assert payload["human_handoff"]["latest_briefs"][0]["target_format"] == "yt_script"
+    assert len(payload["human_handoff"]["fallback_events"]) >= 1
+    assert payload["human_handoff"]["action_candidates"][0]["target_format"] in {"article", "yt_script"}
 
 
 def test_build_editorial_access_indexes_no_data(tmp_path: Path):
@@ -113,3 +121,50 @@ def test_build_editorial_access_indexes_no_data(tmp_path: Path):
     assert payload["digest_at"] is None
     assert payload["status"] == "no-data"
     assert payload["metrics"]["seed_ideas_emitted"] == 0
+    assert payload["human_handoff"]["status"] == "no-data"
+    assert payload["human_handoff"]["latest_briefs"] == []
+
+
+def test_build_editorial_access_indexes_separates_yt_drafts(tmp_path: Path):
+    data = tmp_path / "data"
+    storage = tmp_path / "storage"
+    digest = "20260315T01"
+
+    _write_jsonl(
+        data / "pf_out" / f"pfout_{digest}.jsonl",
+        [{"digest_group_id": f"{digest}::A::Economia::01", "seed_ideas": {"seed_ideas": [{"idea_id": "I1"}]}}],
+    )
+    _write_jsonl(
+        storage / "buses" / "news_piece_brief" / "v1" / "npb_1.jsonl",
+        [{"schema_name": "news_piece_brief.v1", "digest_id_hour": digest, "brief_id": "npb_1", "topic": "Economia"}],
+    )
+    _write_jsonl(
+        data / "drafts" / digest / "article_1.jsonl",
+        [{"schema_name": "news_article_draft.v1", "index_id": "A1", "topic": "Economia", "headline": "H1"}],
+    )
+    _write_jsonl(
+        data / "drafts" / digest / "yt_1.jsonl",
+        [{"schema_name": "news_yt_script_draft.v1", "index_id": "Y1", "topic": "Economia", "headline": "YT1"}],
+    )
+
+    subprocess.run(
+        [
+            sys.executable,
+            str(SCRIPT),
+            "--storage-dir",
+            str(storage),
+            "--data-dir",
+            str(data),
+            "--digest-at",
+            digest,
+        ],
+        capture_output=True,
+        text=True,
+        check=True,
+    )
+
+    payload = json.loads((storage / "indexes" / "editorial_latest.json").read_text(encoding="utf-8"))
+    assert len(payload["human_handoff"]["latest_article_drafts"]) == 1
+    assert len(payload["human_handoff"]["latest_yt_script_drafts"]) == 1
+    assert payload["human_handoff"]["action_candidates"][0]["target_format"] == "yt_script"
+    assert payload["human_handoff"]["action_candidates"][0]["ready_state"] == "draft-ready"
